backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.0
dj-database-url==2.1.0
python-magic==0.4.27
filetype==1.2.0
PyPDF2==3.0.1
pdf2image==1.16.3

trimesh==4.0.5
numpy==1.24.3
pygltflib==1.16.1
PyWavefront==1.3.3

django-cleanup==8.0.0              # Auto-limpia archivos antiguos
validators==0.22.0                 # Validaciones útiles
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
djangorestframework-simplejwt==5.3.1  # JWT Authentication

# Gmail API (OAuth2 email sending — same as n8n Gmail node)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
